backend/requirements.txt
django=3.2
djangorestframework
loguru
psycopg2
django-cors-headers
Pillow